test: divide MessageStream test
